requirements.txt
aiofiles==0.4.0
aiomysql==0.0.20
aiormq==3.2.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.53.2
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
nose==1.3.7
orjson==2.6.3
pamqp==2.3.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
pydantic==1.4
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.2
python-multipart==0.0.5
PyYAML==5.3.1
Random-Word==1.0.4
requests==2.23.0
Rx==1.6.1
six==1.14.0
starlette==0.13.2
traitlets==4.3.3
ujson==2.0.3
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
yarl==1.4.2
